pyproject.toml
[project]
name = "hifisampler"
version = "0.0.6"
description = "A new UTAU resampler based on pc-nsf-hifigan and straycatresampler for virtual singer."
readme = "README.md"
requires-python = ">=3.10, <3.11"
dependencies = [
    "filelock>=3.17.0",
    "librosa>=0.10.2.post1",
    "numba>=0.61",
    "numpy>=2.1.3",
    "onnxruntime>=1.22.1",
    "praat-parselmouth>=0.4.5",
    "pyloudnorm>=0.1.1",
    "pyworld>=0.3.5",
    "pyyaml>=6.0.2",
    "resampy>=0.4.3",
    "ruamel-yaml>=0.18.10",
    "scipy>=1.15.2",
    "soundfile>=0.13.1",
    "torch>=2.7.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128" },
]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
